requirements.txt
ultralytics
torch
streamlit
opencv-python-headless
Pillow 
